+++ pref/lib/src/radio.dart
@@ -5,6 +5,7 @@
 
 import 'package:flutter/material.dart';
 
+import 'log.dart';
 import 'service/pref_service.dart';
 
 class PrefRadio<T> extends StatefulWidget {
@@ -83,8 +84,8 @@ class _PrefRadioState<T> extends State<PrefRadio<T>> {
     T? value;
     try {
       value = PrefService.of(context).get(widget.pref);
-    } catch (e) {
-      print('Unable to load the value: $e');
+    } catch (e, s) {
+      logger.severe('Unable to load the value', e, s);
     }
 
     if (widget.radioFirst) {