+++ pref/lib/src/slider.dart
@@ -7,6 +7,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 
+import 'log.dart';
 import 'service/pref_service.dart';
 
 class PrefSlider<T extends num> extends StatefulWidget {
@@ -103,8 +104,8 @@ class _PrefSliderState<T extends num> extends State<PrefSlider> {
     T? value;
     try {
       value = PrefService.of(context).get<T>(widget.pref);
-    } catch (e) {
-      print('Unable to load the value: $e');
+    } catch (e, s) {
+      logger.severe('Unable to load the value', e, s);
     }
 
     final min = (widget.min ?? 0.0).toDouble();